Header refactor

+++ objc/source/abstract/HandleError.hpp
@@ -29,8 +29,8 @@ namespace WCDB {
 class HandleError : public Error {
 public:
     HandleError();
-    typedef int64_t Tag;
-    static constexpr const Tag invalidTag = INT64_MIN;
+    typedef int32_t Tag;
+    static constexpr const Tag invalidTag = INT32_MIN;
 
     enum Operation {
         NotSet = 0,
